+++ src/main/java/com/mclegoman/perspective/mixin/client/textured_entity/minecraft/end_crystal/EndCrystalEntityRendererMixin.java
@@ -11,7 +11,6 @@ import com.mclegoman.perspective.client.textured_entity.TexturedEntity;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.decoration.EndCrystalEntity;
 import net.minecraft.util.Identifier;
 import org.spongepowered.asm.mixin.Final;
@@ -21,7 +20,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(priority = 10000, value = net.minecraft.client.render.entity.EndCrystalEntityRenderer.class)
 public class EndCrystalEntityRendererMixin {